fix round2_32, split round2 tests because they depend on sizeof int at compile time
