+++ http_test.go
@@ -7,7 +7,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"math"
 	"mime/multipart"
 	"net/http"
 	"net/http/httptest"
@@ -16,7 +15,6 @@ import (
 	"strings"
 	"testing"
 	"time"
-	"unsafe"
 
 	"github.com/valyala/bytebufferpool"
 )
@@ -1965,34 +1963,6 @@ func testSetResponseBodyStreamChunked(t *testing.T, body string, trailer map[str
 		if string(r) != v {
 			t.Fatalf("unexpected trailer %q. Expecting %q. Got %q", k, v, r)
 		}
-	}
-}
-
-func TestRound2ForSliceCap(t *testing.T) {
-	t.Parallel()
-
-	testRound2ForSliceCap(t, 0, 0)
-	testRound2ForSliceCap(t, 1, 1)
-	testRound2ForSliceCap(t, 2, 2)
-	testRound2ForSliceCap(t, 3, 4)
-	testRound2ForSliceCap(t, 4, 4)
-	testRound2ForSliceCap(t, 5, 8)
-	testRound2ForSliceCap(t, 7, 8)
-	testRound2ForSliceCap(t, 8, 8)
-	testRound2ForSliceCap(t, 9, 16)
-	testRound2ForSliceCap(t, 0x10001, 0x20000)
-
-	if unsafe.Sizeof(int(0)) == 4 {
-		testRound2ForSliceCap(t, math.MaxInt32-1, math.MaxInt32)
-	} else {
-		testRound2ForSliceCap(t, math.MaxInt32, math.MaxInt32)
-		testRound2ForSliceCap(t, math.MaxInt64-1, math.MaxInt64-1)
-	}
-}
-
-func testRound2ForSliceCap(t *testing.T, n, expectedRound2 int) {
-	if roundUpForSliceCap(n) != expectedRound2 {
-		t.Fatalf("Unexpected round2(%d)=%d. Expected %d", n, roundUpForSliceCap(n), expectedRound2)
 	}
 }
