+++ round2_32.go
@@ -3,6 +3,8 @@
 
 package fasthttp
 
+import "math"
+
 func roundUpForSliceCap(n int) int {
 	if n <= 0 {
 		return 0
